requirements.txt
astroid==2.2.5
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
Django==2.1.5
django-cors-headers==3.1.0
django-crispy-forms==1.7.2
django-email-utils==1.0.0
django-filter==2.2.0
django-oauth-toolkit==1.2.0
django-pyodbc-azure==2.1.0.0
django-rest-auth==0.9.5
django-rest-email-auth==2.0.2
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
idna==2.8
isort==4.3.17
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysql-connector-python==8.0.15
oauthlib==3.1.0
PyJWT==1.7.1
pylint==2.3.1
pyodbc==4.0.25
pytz==2018.9
requests==2.22.0
six==1.12.0
typed-ast==1.3.1
urllib3==1.25.3
virtualenv==16.4.3
wrapt==1.11.1
